+++ R/utils.R
@@ -209,10 +209,35 @@ printN <- function(x, n=6, collapse=', ') {
     paste0(paste(x, collapse=collapse))
 }
 
-getPamFft <- function(data) {
+# getPamFft <- function(data) {
+#     if(inherits(data, 'PamBinary')) {
+#         # data$data <- contourToFreq(data$data)
+#         return(getPamFft(data$data))
+#     }
+#     if(length(data) == 0) {
+#         return(NULL)
+#     }
+#     if(!(all(c('sliceData', 'nSlices', 'sampleDuration', 'startSample', 'maxFreq') %in%
+#              names(data[[1]])))) {
+#         # stop('Appears data is not a Whistle and Moan Detector binary file.')
+#         return(NULL)
+#     }
+#     tempData <- data[[1]]
+#     if(tempData$sliceData[[1]]$sliceNumber == 0) {
+#         tempData <- data[[2]]
+#     }
+#     fftHop <- (tempData$startSample + 1)/tempData$sliceData[[1]]$sliceNumber
+#     fftLen <- tempData$sampleDuration -
+#         (tempData$sliceData[[tempData$nSlices]]$sliceNumber - tempData$sliceData[[1]]$sliceNumber) * fftHop
+#     sr <- fftLen * tempData$maxFreq /
+#         max(unlist(lapply(tempData$sliceData, function(x) x$peakData)))
+#     list(sr=sr, hop=fftHop, wl=fftLen)
+# }
+
+getPamFft <- function(data, method=c('new', 'old')) {
     if(inherits(data, 'PamBinary')) {
         # data$data <- contourToFreq(data$data)
-        return(getPamFft(data$data))
+        return(getPamFft(data$data, method=method))
     }
     if(length(data) == 0) {
         return(NULL)
@@ -222,22 +247,75 @@ getPamFft <- function(data) {
         # stop('Appears data is not a Whistle and Moan Detector binary file.')
         return(NULL)
     }
-    tempData <- data[[1]]
-    if(tempData$sliceData[[1]]$sliceNumber == 0) {
-        tempData <- data[[2]]
+    if(length(data) == 1) {
+        method <- 'old'
     }
-    fftHop <- (tempData$startSample + 1)/tempData$sliceData[[1]]$sliceNumber
-    fftLen <- tempData$sampleDuration -
-        (tempData$sliceData[[tempData$nSlices]]$sliceNumber - tempData$sliceData[[1]]$sliceNumber) * fftHop
-    sr <- fftLen * tempData$maxFreq /
-        max(unlist(lapply(tempData$sliceData, function(x) x$peakData)))
+    switch(
+        match.arg(method),
+        'new' = {
+            samplePairs <- bind_rows(
+                lapply(data, function(x) {
+                    list(start=x$startSample+1, slice=x$sliceData[[1]]$sliceNumber)
+                }))
+            samplePairs <- distinct(samplePairs)
+            if(nrow(samplePairs) <= 1) {
+                return(getPamFft(data, method='old'))
+            }
+
+            hops <- diff(samplePairs$start) / diff(samplePairs$slice)
+            hops <- unique(hops)
+            hops <- hops[is.finite(hops)]
+            if(length(hops) > 1) {
+                hops <- as.integer(hops)
+                checkPow2 <- round(log2(hops)) == log2(hops)
+                if(!any(checkPow2)) {
+                    fftHop <- hops[1]
+                } else {
+                    fftHop <- which(checkPow2)[1]
+                }
+            } else {
+                fftHop <- hops
+            }
+            fftLen <- data[[1]]$sampleDuration -
+                (data[[1]]$sliceData[[data[[1]]$nSlices]]$sliceNumber - data[[1]]$sliceData[[1]]$sliceNumber) * fftHop
+            sr <- fftLen * data[[1]]$maxFreq /
+                max(unlist(lapply(data[[1]]$sliceData, function(x) x$peakData)))
+            if(any(c(fftHop, fftLen, sr) <= 0) ||
+               any(is.na(c(fftHop, fftLen, sr))) ||
+               any(is.infinite(c(fftHop, fftLen, sr)))) {
+                return(getPamFft(data, method='old'))
+            }
+            # return(list(sr=sr, hop=fftHop, wl=fftLen))
+
+        },
+        'old' = {
+            tempData <- data[[1]]
+            if(tempData$sliceData[[1]]$sliceNumber == 0) {
+                if(length(data) == 1) {
+                    return(NULL)
+                }
+                tempData <- data[[2]]
+            }
+            fftHop <- (tempData$startSample + 1)/tempData$sliceData[[1]]$sliceNumber
+            fftLen <- tempData$sampleDuration -
+                (tempData$sliceData[[tempData$nSlices]]$sliceNumber - tempData$sliceData[[1]]$sliceNumber) * fftHop
+            sr <- fftLen * tempData$maxFreq /
+                max(unlist(lapply(tempData$sliceData, function(x) x$peakData)))
+            # return(list(sr=sr, hop=fftHop, wl=fftLen))
+            if(any(c(fftHop, fftLen, sr) <= 0) ||
+               any(is.na(c(fftHop, fftLen, sr))) ||
+               any(is.infinite(c(fftHop, fftLen, sr)))) {
+                return(NULL)
+            }
+        }
+    )
     list(sr=sr, hop=fftHop, wl=fftLen)
 }
 
 ppVars <- function() {
     list(nonModelVars = c('UID', 'Id', 'parentUID', 'sampleRate', 'Channel',
                           'angle', 'angleError', 'peakTime', 'depth', 'sr',
-                          'annoId', 'inAnno', 'db'),
+                          'annoId', 'inAnno', 'db', 'gpsUncertainty'),
          # tarMoCols = c(
          #     "TMModelName1", "TMLatitude1", "TMLongitude1", "BeamLatitude1",
          #     "BeamLongitude1", "BeamTime1", "TMSide1", "TMChi21", "TMAIC1", "TMProbability1",
@@ -262,6 +340,8 @@ ppVars <- function() {
                                      'MmMe', 'MmMe', 'MmMe')),
          dglCols = c('Id', 'UID', 'UTC', 'UTCMilliseconds', 'PCLocalTime', 'PCTime',
                      'ChannelBitmap', 'SequenceBitmap', 'EndTime', 'DataCount'),
+         ctCols = c('Chi2', 'median_IDI_sec', 'mean_IDI_sec', 'std_IDI_sec', 'algorithm_info',
+                     'avrg_spectrum_max', 'avrg_spectrum', 'classifiers', 'speciesFlag'),
          binPattern = '(Clicks|WhistlesMoans|GPL).*pgdf$'
     )
 }
@@ -424,14 +504,15 @@ getTimeRange <- function(x, mode=c('event', 'detection'), sample=FALSE) {
                    nrow(d) == 0) {
                     return(NULL)
                 }
-                out <- d[, c('UID', 'UTC', 'duration'), drop = FALSE]
+                out <- d[, c('UID', 'UTC'), drop = FALSE]
                 if('duration' %in% colnames(d)) {
                     switch(attr(d, 'calltype'),
                            'whistle' = out$duration <- d$duration,
                            # 'click' = out$duration <- d$duration / 1e6,
                            'click' = out$duration <- 0, # duration is not reliable for clicks
                            'cepstrum' = out$duration <- d$duration,
-                           'gpl' = out$duration <- d$duration
+                           'gpl' = out$duration <- d$duration,
+                           out$duration <- 0
                     )
                 } else {
                     out$duration <- 0